Add files via upload
It is tracking application which uses wifi ap's and not GPS. First you have start xampp for SQL connectivity. Open this project in eclipse and run the login page via server. Login page will appear here user will login in user mode and driver will login in driver mode. User will go to homepage where he will enter the source and destination he requires and then we are also giving reminder system for user i.e. if he want to get us remind on the station which he wants to. On driver side, he can add wifi manually on the station by scanning the wifi ap's the strongest network will get added in database on the selected station. To collect database of bus for routes and wifi ap's we have done training of application where we are adding wifi ap's and route on database based on their longitude & latitude of that wifi.